src/backend/test_files/test_qdrant_method.py
from qdrant_client import QdrantClient
from dotenv import load_dotenv
import os

load_dotenv()

client = QdrantClient(
    url=os.getenv("QDRANT_URL"),
    api_key=os.getenv("QDRANT_API_KEY"),
    prefer_grpc=False   # only keep this
)

print("Available methods:")
print([m for m in dir(client) if "search" in m.lower()])